requirements.txt
bcrypt == 3.2.0
email-validator == 1.1.3
Flask == 2.0.1
Flask-Bcrypt == 0.7.1
Flask-Login == 0.5.0
Flask-Mail == 0.9.1
Flask-SQLAlchemy == 2.5.1
Flask-WTF == 0.15.1
Jinja2 == 3.0.1
Pillow == 8.3.1
python-dotenv == 0.19.0
SQLAlchemy == 1.4.21
WTForms == 2.3.3
psycopg2
gunicorn
python-dotenv